Octopode interactions with ring damage bonuses
Octopodes hold each weapon in a pair of tentacles.  This implies that
they also hit bare-handed with a pair of tentacles, and should get
damage bonuses from rings worn on both tentacles used.  Now they do.

- Rings 0 and 2 (shown as 1 and 3) are left-hand rings
- Rings 1 and 3 (shown as 2 and 4) are right-hand rings

This applies to sear damage, poison, and the +5 damage bonus from
hitting with an EDDER_SYMBOL ring.  Artifact rings are unaffected and
always apply regardless of the tentacle they are worn on, as before.

The messaging for sear damage is a bit weird - it
includes (deduplicated) properties from both rings as if it was a
single ring, except that "ring" is pluralized.  For example, if you're
wearing a jade ring and a silver ring and have Eden bound:

 Your silver skin and jade silver rings sear the pit fiend's flesh!

Also fix some variables being used uninitialized.