Bring back force armor from ndnh1
3 AC, 3 DR, MC 3, heavy armor made of gemstone, future tech
Ana hlf starts with it
Knocks back and stuns attackers

Unlike in ndnh1, it is implemented mvm and uvm.  When worn by a
monster, it basically makes it unfightable in melee unless you have
stun resistance and are at least huge.  This isn't really a problem
since the only way to get it is to start as ana hlf or wish for it as
a tourist.