Remove wornmask gaps and don't give intrinsic extrinsics
A side-effect of this is that wornmasks now fit into 32 bits again,
although I didn't change them back to long.

Acu now has HTelepat (special-cased to never lose it and have it act
as Unblind_telepat) instead of intrinsic ETelepat.  Extrinsics should
never be given as intrinsics because the wornmasks conflict.