Require a standard C compiler and remove MACRO_CPATH, VISION_TABLES
A standard C compiler (NHSTDC) is assumed.

The macro version of clear_path (MACRO_CPATH) is removed: it's
redundant and the compiler can inline it if necessary.

VISION_TABLES is removed: it's unnecessary (calculating it at runtime
is fine) and nobody tested with it or used it.

VOID_ARGS is removed: void is now used everywhere in prototypes.

MACsansMPWTOOL is removed: I think that's some system-specific define
that I missed.