User sounds support using libsndfile and portaudio
fork()s before playing the sound so it doesn't block the game.  If a
sound is already playing, kills the old process (by storing its pid in
nowplayingpid and registering a handler for SIGCHLD to unset it when
it's killed).  Also registers an atexit() handler so sound stops
playing if you exit the game.

Maybe this should just exec something to play the sound instead of
doing it with a library?

No volume support yet.