Bugfix: curses command count was capped at LARGEST_INT (32767)
This made it impossible to repeat a single command more than 32767
times, or take out/put in a partial stack of more than 32767 items.
It's now capped at INT_MAX (2147483647 with 32-bit int) instead and
uses C23 stdckdint.h macros to avoid signed integer overflows (which
are undefined behaviour).  If stdckdint.h is unavailable (it was only
added in GCC 14), uses the GCC __builtin_*_overflow functions, which
are supported since GCC 5.  If these too are unavailable, falls back
to an implementation using C11 _Generic.  The fallback ckd_mul
implementation just casts to a long long, so won't work for anything
larger than a 32-bit integer (assuming 64-bit long long), which is
fine for now since nothing since it's only used on ints.  I'm not sure
the fallback implementations are even necessary, since we already
require either C23 or a GCC extension (__VA_OPT__) for macromagic.