Remove doc/tmac.n because it's non-free
"This package may be circulated freely with the news documentation; it
may not be sold, but is to be distributed with the unformatted news
documents."

"Further, please do not distribute this package if you make any
changes"

I think this is used for the guidebook in vanilla, but not actually
used by the *dNetHack build system.  A free replacement needs to be
found.